FEM: selection observer, move class into selection widgets module
