--- src/Mod/Fem/femguiobjects/FemSelectionObserver.py
@@ -1,47 +1,0 @@
-# ***************************************************************************
-# *                                                                         *
-# *   Copyright (c) 2015 - Bernd Hahnebach <bernd@bimstatik.org>            *
-# *                                                                         *
-# *   This program is free software; you can redistribute it and/or modify  *
-# *   it under the terms of the GNU Lesser General Public License (LGPL)    *
-# *   as published by the Free Software Foundation; either version 2 of     *
-# *   the License, or (at your option) any later version.                   *
-# *   for detail see the LICENCE text file.                                 *
-# *                                                                         *
-# *   This program is distributed in the hope that it will be useful,       *
-# *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
-# *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
-# *   GNU Library General Public License for more details.                  *
-# *                                                                         *
-# *   You should have received a copy of the GNU Library General Public     *
-# *   License along with this program; if not, write to the Free Software   *
-# *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
-# *   USA                                                                   *
-# *                                                                         *
-# ***************************************************************************
-
-__title__ = "Selection Observer"
-__author__ = "Bernd Hahnebach"
-__url__ = "http://www.freecadweb.org"
-
-## \addtogroup FEM
-#  @{
-
-import FreeCAD
-import FreeCADGui
-
-
-class FemSelectionObserver:
-    '''FemSelectionObserver'''
-    def __init__(self, parseSelectionFunction, print_message=''):
-        self.parseSelectionFunction = parseSelectionFunction
-        FreeCADGui.Selection.addObserver(self)
-        FreeCAD.Console.PrintMessage(print_message + "!\n")
-
-    def addSelection(self, docName, objName, sub, pos):
-        selected_object = FreeCAD.getDocument(docName).getObject(objName)  # get the obj objName
-        self.added_obj = (selected_object, sub)
-        # on double click on a vertex of a solid sub is None and obj is the solid
-        self.parseSelectionFunction(self.added_obj)
-
-##  @}
